+++ src/botalk.json
@@ -10,7 +10,7 @@
 },
 {
     "id": 3,
-    "desc": "Kiitos. Nyt aloitamme. Missä haluaisit tehdä pestisi Kajolla?",
+    "desc": "Kiitos. Aloittakaamme. Missä haluaisit tehdä pestisi Kajolla?",
     "type": "where"
 },
 {
